Cambios de Daniel

+++ API Gateway/Gateway.Aplicacion/Gateway.Aplicacion/Alquiler/Request/RegistrarAlquilerRequest.cs
@@ -8,9 +8,9 @@ namespace Gateway.Aplicacion.Alquileres.Request
 {
     public class RegistrarAlquilerRequest
     {
-        public int IdAlquiler { get; set; }
-        public int IdPelicula { get; set; }
-        public int IdUsuario{ get; set; }
+        public double IdAlquiler { get; set; }
+        public double IdPelicula { get; set; }
+        public double IdUsuario{ get; set; }
         public string FecInicio { get; set; }
         public string FecFin { get; set; }
     }
